feat(US-001): extract manifest types to manifest-types.ts
